+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNPB0DJ8E4GX9SEE2CQ59Q93"
+        "01KNPFA4WSG0T7B3ZSCG5XVSMV"
     )
     r"""The unique identifier of the data source"""