+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNPB0DJ6QTR6EYEHMS18BJX2"
+        "01KNPFA4WP7MPNKT75PR8R2RWQ"
     )
     r"""The unique identifier of the data source"""