+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNPB0DJA3FWMXNWX8G192P7N"
+        "01KNPFA4WV5P9F5CPVP9V7A6WV"
     )
     r"""The unique identifier of the data source"""