+++ packages/orq-rc/RELEASES.md
@@ -6939,3 +6939,13 @@ Based on:
 - [python v4.7.0-rc.54] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.54] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.54 - packages/orq-rc
+
+## 2026-04-08 12:01:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.55] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.55] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.55 - packages/orq-rc